test: 토큰 설정 및 페이지 개
