+++ src/test/java/roomescape/controller/BaseControllerTest.java
@@ -2,19 +2,24 @@ package roomescape.controller;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
+import org.springframework.boot.test.mock.mockito.SpyBean;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.annotation.DirtiesContext.ClassMode;
 import roomescape.config.TestConfig;
+import roomescape.security.AuthArgumentResolver;
+import roomescape.security.AuthInterceptor;
+import roomescape.security.JwtTokenProvider;
 
 @SpringBootTest(
         classes = TestConfig.class,
         webEnvironment = WebEnvironment.RANDOM_PORT
 )
 @DirtiesContext(classMode = ClassMode.BEFORE_EACH_TEST_METHOD)
-abstract class BaseControllerTest {
+public abstract class BaseControllerTest {
 
     @LocalServerPort
     private int port;
