+++ src/test/java/roomescape/controller/web/MemberPageControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.web;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -8,11 +8,17 @@ import io.restassured.response.Response;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.http.HttpStatus;
+import roomescape.controller.BaseControllerTest;
 
-class PageControllerTest extends BaseControllerTest {
+class MemberPageControllerTest extends BaseControllerTest {
 
-    @ParameterizedTest(name = "{0} 페이지를 조회한다.}")
-    @ValueSource(strings = {"/", "/reservation", "/admin", "/admin/reservation", "/admin/time", "/admin/theme"})
+    @ParameterizedTest(name = "{0} 페이지를 조회한다.")
+    @ValueSource(strings = {
+            "/",
+            "/reservation",
+            "/login",
+            "/signup"
+    })
     void pageTest(String path) {
         ExtractableResponse<Response> response = RestAssured.given().log().all()
                 .when().get(path)
@@ -22,4 +28,3 @@ class PageControllerTest extends BaseControllerTest {
         assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
     }
 }
-
