+++ src/test/resources/application.yml
@@ -10,3 +10,8 @@ logging:
   level:
     org.springframework:
       jdbc: DEBUG
+
+security:
+  jwt:
+    secret-key: abcdefghijklmnopqrstuvwxyz0123456789
+    expiration-time: 3600000 # 1 hour
